net/mlx5_core: Introduce crdump utility for FW debug
Add the ability to collect the crdump area in the pci whenever there is
a catastrophic condition.

That data is collected and can be shown via proc entry. A print in
the message log will show the path to the crdump file.

The driver collects the crdump on the first time it is needed, under the
assumption that the first time is the root cause of all the others.

Signed-off-by: Erez Shitrit <erezsh@mellanox.com>
Signed-off-by: Alex Vesker <valex@mellanox.com>
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>