net/mlx5: Issue SW reset on FW assert
If a FW assert is considered fatal, indicated by a new bit in the health
buffer, reset the FW. After the reset go through the normal recovery
flow. Only one PF needs to issue the reset, so an attempt is made to prevent
the 2nd function from also issuing the reset. It's not an error if that
happens, it's just slows recovery.

issue: 1187431
Change-Id: If80e852f3e5e8ea69311b16b48f3acebfddee7c2
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>