{net,IB}/mlx5: Raise fatal IB event when sys error occurs
All other mlx5_events report the port number as 1 based, which is how FW
reports it in the port event EQE. Reporting 0 for this event causes
mlx5_ib to not raise a fatal event notification to registered clients
due to a seemingly invalid port.

All switch cases in mlx5_ib_event that go through the port check are
supposed to set the port now, so just do it once at variable
declaration.

issue: 1272952
fixes: 89d44f0a6c73("net/mlx5_core: Add pci error handlers to mlx5_core
driver")
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>

Change-Id: If897bac93a8d2e327df485cba3ae5f86373b982c